moving test and fixing issue
