+++ conans/test/functional/python_requires/python_requires_test.py
@@ -826,49 +826,3 @@ class MyConanfileBase(ConanFile):
         client.run("package .")
         self.assertIn("conanfile.py: Pkg1 package: 42", client.out)
         client.run("export-pkg . pkg1/0.1@user/testing")
-
-    def test_transitive_python_requires(self):
-        # https://github.com/conan-io/conan/issues/8546
-        client = TestClient()
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile
-
-            myvar = 123
-
-            def myfunct():
-                return 234
-
-            class SharedFunction(ConanFile):
-                name = "shared-function"
-                version = "1.0"
-            """)
-        client.save({"conanfile.py": conanfile})
-        client.run("export . @user/channel")
-
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile
-
-            class BaseClass(ConanFile):
-                name = "base-class"
-                version = "1.0"
-                python_requires = "shared-function/1.0@user/channel"
-
-                def build(self):
-                    v = self.python_requires["shared-function"].module.myvar  # v will be 123
-                    f = self.python_requires["shared-function"].module.myfunct()  # f will be 234
-                    self.output.info("%s, %s" % (v, f))
-            """)
-        client.save({"conanfile.py": conanfile})
-        client.run("export . user/channel")
-
-        conanfile = textwrap.dedent("""
-            from conans import ConanFile
-
-            class Consumer(ConanFile):
-                name = "consumer"
-                version = "1.0"
-                python_requires = "base-class/1.0@user/channel"
-                python_requires_extend = "base-class.BaseClass"
-            """)
-        client.save({"conanfile.py": conanfile})
-        client.run("create . @user/channel")
